+++ src/Framework/Storage/StoragePage.cpp
@@ -10,7 +10,7 @@ void StoragePage::SetOffset(uint64_t offset_)
     offset = offset_;
 }
 
-uint32_t StoragePage::GetOffset()
+uint64_t StoragePage::GetOffset()
 {
     return offset;
 }
